+++ src/Screens/DictionaryPage.test.js
