+++ src/Screens/TacticsPage.test.js
