+++ src/Screens/TrainingPage.test.js
